[#10] Add a guard_error to the workflow when attempting invalid events
When calling `#may_event?`, `#event`, or `#event!`, if the event is not
supported by the current state, add `:invalid_event` to the `guard_errors`.

I considered `raise NoMethodError.new("#{method_name} is an invalid
transition from #{current_state.name}”)`, but has the following downsides:

1. `may_event?` should not raise an error because it's a basic predicate
and in my opinion should simply return `false` if it can't do what you're
asking. Changing the code to maintain this behavior while raising an
error on the transition would require significant refactoring.

2. I don’t like raising exceptions when they’re likely to be triggered by
user actions, but fully support them when triggered by developer mistakes.
Especially in the case of `may_event?`, I think this error could lead to
user-visible errors (i.e. 500 errors).

3. Raising an error on this would be a breaking change. There are likely
few applications using this gem and it's a 0.X release, but it's still
a concern. I did my applications' specs against the `raise` version and
while everything was green, I did see the following issue in practice (we
use current state to decide which next states to show buttons for instead
of using `may_event?`).

3. Considering the case where this is used in a web application: I have
two tabs of the same object with a button to transition to the next state.
On the first tab, I click the button and it transitions successfully. I switch
to the second tab and click the transition button. With the `raise` version
I get a 500 error. This feels like a bad user experience. The controller action
is essentially (with some `before_action`s and I18n):

    ```
  def publish
    workflow = FlowMachine::Workflow.for(resource).publish!(current_user: current_user)
    if workflow.publish!
      redirect_to resource, notice: "The post was published"
    else
      flash[:now] = "The post could not be published"
      render :show
    end
    ```

  If messaging is important in your case, the `else` case could be changed to
  have a reason based on the `workflow.guard_errors`. This method also prevents
  a similar multi-tab case of one tab changing things that make a `guard` method
  (or `ActiveModel` validation) invalid. We don't do anything fancy, but you
  could easily change the `flash` message to be dynamic based on the `guard_errors`
  or `ActiveModel#errors`.